+++ src/ideabox/policy/content/state_progress.py
@@ -25,5 +25,5 @@ class StateProgress(Container):
 class StateProgressView(BrowserView):
     def get_state_progress(self):
         return api.content.find(
-            contect=self, portal_type="state_progress", sort_on="state_date"
+            context=self.context, portal_type="state_progress", sort_on="state_date"
         )