+++ src/ideabox/policy/content/priority_action.py
@@ -1,10 +1,12 @@
 # -*- coding: utf-8 -*-
+from collective.z3cform.select2.widget.widget import MultiSelect2FieldWidget
 from ideabox.policy import _
 from ideabox.policy.content.project import IProject
 from ideabox.policy.content.project import Project
 from ideabox.policy.content.project import ProjectView
 from plone import api
 from plone.app.textfield.value import IRichTextValue
+from plone.autoform import directives as form
 from plone.indexer.decorator import indexer
 from Products.CMFPlone.utils import getToolByName
 from zope import schema
@@ -13,6 +15,17 @@ from zope.interface import implementer
 
 class IPriorityAction(IProject):
     """IPriorityActionIProject"""
+
+    form.widget(project_district=MultiSelect2FieldWidget)
+    form.order_after(project_district='project_theme')
+    project_district = schema.List(
+        title=_(u"District(s) concerned"),
+        value_type=schema.Choice(
+            title=_(u"District(s) concerned"),
+            vocabulary=u"collective.taxonomy.district",
+        ),
+        required=False,
+    )
 
     strategic_objectives = schema.TextLine(
         title=_(u"Strategic Objectives"), required=False